Move files to be under a feature folder
