+++ src/features/shell-sync/main/compute-unix-shell-environment.global-override-for-injectable.ts
