+++ src/features/shell-sync/main/compute-unix-shell-environment.test.ts
@@ -7,10 +7,10 @@ import type { DiContainer } from "@ogre-tools/injectable";
 import type { ChildProcessWithoutNullStreams } from "child_process";
 import EventEmitter from "events";
 import { flushPromises } from "../../../common/test-utils/flush-promises";
-import type { Spawn } from "../../child-process/spawn.injectable";
-import spawnInjectable from "../../child-process/spawn.injectable";
-import randomUUIDInjectable from "../../crypto/random-uuid.injectable";
-import { getDiForUnitTesting } from "../../getDiForUnitTesting";
+import type { Spawn } from "../../../main/child-process/spawn.injectable";
+import spawnInjectable from "../../../main/child-process/spawn.injectable";
+import randomUUIDInjectable from "../../../main/crypto/random-uuid.injectable";
+import { getDiForUnitTesting } from "../../../main/getDiForUnitTesting";
 import type { ComputeUnixShellEnvironment } from "./compute-unix-shell-environment.injectable";
 import computeUnixShellEnvironmentInjectable from "./compute-unix-shell-environment.injectable";
 import processEnvInjectable from "./env.injectable";
