+++ src/features/shell-sync/main/env.injectable.ts
