+++ src/features/shell-sync/main/execPath.injectable.ts
